Declare Flo and Deb to be devices shipping in JB.
New devices in N and above should declare that they are new devices to
avoid shipping a real WebView (instead we ship a 'stub', a small
replacement apk).

Bug: 26381401
Change-Id: I292d33e2422f74999bbae38a4043030fa5e383f1

flox: Set shipping API level in the common product makefile

It's not meant to be set as a board config flag. Simply do it by
inheriting the makefile that now also exists for devices launched
with Jelly Bean.

Change-Id: I2ca1f9d207be90c58495b3e20c5e2275e09b18bc